+++ CHANGELOG.md
@@ -40,7 +40,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The help popup now lists the keybindings next to what they do rather than in
   front of it
 - Pressing `s` on the working copy now offers to squash into the parent (when there is exactly one)
-- In log tab, the details panel update no longer blocks the UI thread.
+- In log and bookmarks tab, the details panel update no longer blocks the UI
+  thread, and the changes it has shown are cached
 
 ### Fixed